requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
dataclasses-json==0.5.7
diffusers==0.14.0
ffmpeg-python==0.2.0
filelock==3.10.0
Flask==2.2.3
Flask-RESTful==0.3.9
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.116
lit==15.0.7
llvmlite==0.39.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.2
openai-whisper==20230314
packaging==23.0
Pillow==9.4.0
pycodestyle==2.10.0
pydantic==1.10.6
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
sympy==1.11.1
tenacity==8.2.2
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
